warehouse_quote_app/app/core/auth/jwt.py
"""
JWT token handling utilities.
"""

from datetime import datetime, timedelta
from typing import Optional, Dict, Any
from jose import jwt, JWTError

from warehouse_quote_app.app.core.config import settings

def create_access_token(
    data: Dict[str, Any],
    expires_delta: Optional[timedelta] = None
) -> str:
    """Create JWT access token."""
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(
        to_encode,
        settings.SECRET_KEY,
        algorithm=settings.ALGORITHM
    )
    return encoded_jwt

def decode_access_token(token: str) -> Dict[str, Any]:
    """Decode JWT access token."""
    return jwt.decode(
        token,
        settings.SECRET_KEY,
        algorithms=[settings.ALGORITHM]
    )

def get_token_payload(token: str) -> Optional[Dict[str, Any]]:
    """Get payload from JWT token."""
    try:
        return decode_access_token(token)
    except JWTError:
        return None

# Export jose.jwt functions to maintain compatibility
encode = jwt.encode
decode = jwt.decode
